transformers/requirements.txt
git+https://github.com/huggingface/transformers
datasets
psutils
bentoml>=1.0.0a3
